Un-refactor Makefile, sfinae.h, and a bit more :-)
